requirements.txt
asgiref==3.8.1
Django==4.2.17
django-ckeditor==6.7.1
django-filter==24.2
django-jazzmin==2.6.1
django-js-asset==2.2.0
django-modeltranslation==0.19.3
pillow==10.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
